% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_haplotype_ids_to_haplotypes.R
\name{convert_haplotype_ids_to_haplotypes}
\alias{convert_haplotype_ids_to_haplotypes}
\title{Convert haplotype IDs to haplotypes}
\usage{
convert_haplotype_ids_to_haplotypes(haplotype_ids)
}
\description{
Convert one or more haplotype IDs to haplotypes,
in MHCnuggets naming format
}
\examples{
convert_haplotype_ids_to_haplotypes("h13")
}
\author{
Richèl J.C. Bilderbeek
}
