% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_haplotypes_to_haplotype_ids.R
\name{convert_haplotypes_to_haplotype_ids}
\alias{convert_haplotypes_to_haplotype_ids}
\title{Convert haplotypes to haplotype IDs}
\usage{
convert_haplotypes_to_haplotype_ids(haplotypes)
}
\description{
Convert one or more haplotypes (in MHCnuggets naming format)
to haplotype IDs
}
\examples{
convert_haplotype_to_haplotype_ids("HLA-A33:01")
}
\author{
Richèl J.C. Bilderbeek
}
