% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ic50_threshold.R
\name{get_ic50_threshold}
\alias{get_ic50_threshold}
\title{Get the IC50 value to serve as a threshold,
for example, get the IC50 value from below which the
peptides bind strongest}
\usage{
get_ic50_threshold(n_aas = 9, mhc_haplotype, percentile = 0.02)
}
\arguments{
\item{n_aas}{number of amino acids}

\item{mhc_haplotype}{haplotype of either HMC-I or MHC-II.
Use \link{get_mhc1_haplotypes} to get a list of all MHC-I haplotypes.
Use \link{get_mhc2_haplotypes} to get a list of all MHC-II haplotypes.}

\item{percentile}{the IC50 percentile. The IC50 that separates the top
\code{percentile} from the worse binders}
}
\description{
Get the IC50 value to serve as a threshold,
for example, get the IC50 value from below which the
peptides bind strongest
}
