% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc2_haplotypes.R
\name{get_mhc2_haplotypes}
\alias{get_mhc2_haplotypes}
\title{Get the MHC-II haplotypes used in this study}
\usage{
get_mhc2_haplotypes()
}
\description{
Get the MHC-II haplotypes used in this study
}
\examples{
library(testthat)

mhc2_haplotypes <- get_mhc2_haplotypes()
expect_true("HLA-DRB1*0101" \%in\% mhc2_haplotypes)
expect_true("HLA-DQA1*0501/DQB1*0201" \%in\% mhc2_haplotypes)
}
\author{
Richèl J.C. Bilderbeek
}
