% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  binding_strength_threshold,
  fasta_filename,
  haplotypes,
  ic50_threshold,
  max_hydrophobicity,
  mhc_haplotype,
  mhc_1_haplotype,
  mhc_2_haplotype,
  min_hydrophobicity,
  n_aas,
  n_peptides,
  peptide_length,
  protein_sequence,
  protein_sequences,
  source,
  verbose
)
}
\arguments{
\item{binding_strength_threshold}{threshold of the MHC2 allele's
binding strength. A low threshold, e.g. \code{5.0}, selects for the top
5 percent best binders}

\item{fasta_filename}{name of a FASTA file}

\item{haplotypes}{a list of MHC-I or MHC-II haplotypes}

\item{ic50_threshold}{peptides with a predicted IC50 value (in nM)
lower than this will be considered binders.
A threshold of 500 (nM) is common.}

\item{max_hydrophobicity}{the maximal hydrophobicity}

\item{mhc_haplotype}{haplotype of either HMC-I or MHC-II.
Use \link{get_mhc1_haplotypes} to get a list of all MHC-I haplotypes.
Use \link{get_mhc2_haplotypes} to get a list of all MHC-II haplotypes.}

\item{mhc_1_haplotype}{HMC-I haplotope.
Use \link{get_mhc1_haplotypes} to get a list of all haplotypes}

\item{mhc_2_haplotype}{HMC-II haplotope.
Use \link{get_mhc2_haplotypes} to get a list of all haplotypes}

\item{min_hydrophobicity}{the minimal hydrophobicity}

\item{n_aas}{number of amino acids}

\item{n_peptides}{number of peptides}

\item{peptide_length}{length of the peptide in amino acids}

\item{protein_sequence}{a protein sequence}

\item{protein_sequences}{one or more protein sequences}

\item{source}{source of IC50s:\cr
\enumerate{
  \item \code{random} base quantiles on randomly simulated peptides
  \item \code{NA} use old interface, to become obsoleted
}}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
